add connection status and rename sec to cvars
