+++ integrations/google/utils/conn_status.go
@@ -1,4 +1,4 @@
-package google
+package utils
 
 import (
 	"context"
